requirements.txt
xlrd==1.2.0
pandas==1.2.4
requests==2.25.1
numpy==1.20.3
beautifulsoup4==4.9.3
Pillow==8.2.0
